pyproject.toml
[project]
name = "free"
version = "0.1.0"
description = "Churn prediction ML pipeline with FastAPI and Streamlit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datetime>=6.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "seaborn>=0.13.2",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "streamlit>=1.28.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "pyarrow>=14.0.0",
]
